A hydra-themed ghost vault (mikee)
In gammafunk_ghost_hydra_chop, a fallen player made the mistake of using
a non-flaming edged weapon vs a hydra, leaving their ghost, the weapon,
and of course a hydra with more heads than usual. The weapon was not the
fallen player's actual weapon, but we can allow a little creative
fiction here, especially if the player was a felid!. We place only
weapon types usable by all non-felid species and always use a
non-flaming ego. There is a 1/3 chance for the weapon to be artefact.

The vault can also place a limited amount of additional loot and extra
hydras, scaling with depth, with the max possible hydra heads going as
high as 16 in late Swamp. Also places in D:8- and Lair.