Make sticks to snakes and sandblast cost 10x as many missiles
We increased the spawn rate of stones and arrows by a factor of 8, which
gave these spells effectively infinite ammo. These spells are already
really good, and the ammo cost is all there is to go against them.

I am overcompensating here by making the cost ten arrows/stones, because
these spells could do with a nerf anyway, and 10 is a nicer number than
8.

In order to compensate for this, Tm and EE now start with 10x as many
missiles.