Don't assume elemental force placement doesn't change clouds
This code assumed that the values returned by `cloud_at` were unchanged
from the first pass to all points of the second pass. However, in the
abyss, placement of elementals was also adding tloc clouds, which are
not in the map, and therefore `*map_find` was not safe. This could cause
crashes, which were relatively reliably triggered by reading summoning,
fog, and triggering elemental force in succession. (I think the impact
of summoning was just to get a crowd of already placed monsters around,
though I didn't test carefully why this mattered.)