Improvments and fixes to a ghost vault
In gammafunk_ghost_hydra_chop, fix the weights so that they're scaled
enough to prevent a weight of 0, properly exclude no ego from
artefact weapons, and fix the antimagic ego specifier. Add holy wrath
into the possible weapon egos, taking care not to try to apply it to
demon weapons.

Also move the weapon producing code to its own function, and clean up
the vegetation placement in the vault.