Remove amulets of harm
These amulets were generally seen as bad by the entire Crawl playerbase.
There is something to be said for nerfing the effects of harm, but
generally it is a poor mechanic that isn't very transparent and never
gets used because people would rather take less damage most of the time.

This item was almost exclusively used because it had *Drain, which can
be very punishing early on for some characters, and this item alone
means that it is optimal to identify amulets in the early game (and
beyond).

Hence, by a vote of 15 to 8, with 2 wanting a rework of this amulet, we
remove amulets of harm.