A ghost vault themed after Okawaru's wrath (hellmonk)
In gammafunk_ghost_fury_of_okawaru, the player finds a ghost on the
altar of Okawaru along with a bunch of representative Oka wrath monsters
and some loot. Although the generated ghost likely wasn't the result of
dying to "the Fury of Okawaru", hopefully this vault evokes the common
occurrence of players trying and failing to switch from Oka to another
god. It places in Dungeon, Orc, Vaults, Elf, and Crypt, and uses
warrior, archer, and giant monsters from Oka's wrath set. The Vaults
variant place through a separate vault room.

Original idea courtesy of hellmonk. Add new and exciting food costs.