Extend the Vaults layout to place ghost vault rooms
Vault rooms placing ghosts should now set a `vaults_ghost` tag to be
selected as part of the chance for each level to have a ghost vault. The
Vaults layout rolls based on the chance value in dat/dlua/ghost.lua and
selects from rooms with the vaults_ghost tag. Vault rooms that place
ghosts should call the function vaults_ghost_setup() to set this and
other tags we want all ghost vaults to have.