Format, unmeme, and add full species descript for Hermit Crabs
Some formatting fixes, unmemeing front-facing text (debugging still is
full of memes), and adding a full species description for Hermit Crabs
in the Crawl manual. I also made it so they can't be deformed to fit
with the other drastic body mutations (I left the Jiyva ones alone for
the moment).