Rename Tzitzimimeh to Shadow Fiends
There are two reasons why Tzitzimitl is a bad name for this monster:
firstly, it is unpronounceable which inhibits communication and memory
for new players of what this monster does, and secondly it does not
convey the same consistent meaning of "monster that torments and does
something else" that Fiend does. Additionally, naming monsters after
ancient Aztec deities is not interesting. This commit closes #34.