Implement spooky vaults, part 2
This is a squashed version of the last fifteen commits from advil and
gammafunk, created from the merging of crawl/crawl:spooky_vaults into
crawl/crawl:master :

 b611b69 Make bones file resolution coarser (first pass)
 9e38d4b Reformat vault guidelines
 30c74a2 Ghost vault functions and ghost DES file cleanup
 99b25f0 Extend the Vaults layout to place ghost rooms
 a2df02c Update ghost vault guidelines
 c69a6cf Some ghost vaults for the Vaults branch
 6d380fd A set of general ghost vaults
 8f30707 Some thematic ghost vaults
 28d1087 A ghost vault that places multiple ghosts
 c41e3e3 A hydra-themed ghost vault (mikee)
 84fad1d A ghost vault themed after Okawaru's wrath (hellmonk)
 1d9e83a TAG MAJOR comment for Forest/Blade branch entries
 7f357de Remove an unused error tracking variable
 0943f25 Use the correct integer variable type.

This completes the merging of spooky vaults into X-Crawl. A massive
thanks to advil and gammafunk for pushing through with this quickly in
the end, and solving the five-year-old problems of ghosts being largely
annoying and non-threatening (particularly with respect to melee-only
ghosts), but sometimes being able to cast incredibly strong spells (for
example fireball/lightning bolt ghosts on D:6, and level 9 spells
eventually too), with the player being able to do nothing about it.
Additionally, ghosts were fundamentally flawed in design by being fixed
to a floor, because this meant players could repeatably abuse stairs to
deal with them. Ghosts also had issues with being noticeably more
frequent in the early game, and the potential for ghost griefing from
players (as well as bad players which may be accused of ghost griefing
even though it is simply their incompetence).

These thirty commits solve the issues with ghosts and can be summarised
as follows:
  - Ghosts no longer spawn randomly on floors;
  - Ghosts now spawn roughly 10% of the time in specially placed ghost
    vaults, with the ghost(s) clearly visible behind transparent walls
    and trapped behind either a runed door or transporter;
  - Some of these ghost vaults will contain other thematic monsters as
    as well as the ghost, and one contains multiple ghosts;
  - There are special provisions for placing ghost vaults in the Vaults
    branch, since all rooms in Vaults are specially crafted vaults; and
  - Ghosts can now use stairs.