Adjust the vault muffindrake_ghost_dual_entry
Rename this vault to muffindrake_ghost_dual_entry to be in line with our
other vault names. Add windows for the back of the vault and a dungeon
connector so that the player will always be able to inspect it and find
any ghost, if they so wish. Also down-weight the vault in its earlier
depths of D:10-11. Among ghost vaults it should be a less commonly seen
challenge, since it places two 8, a 9, and a lich, which is (almost)
guaranteed to overshadow the ghost itself.

Randomize the placement of the 1, 2, and 3 monsters, also randomizing
their placement numbers slightly but with the same average numbers.
Slightly increase loot quality by giving another superb_item instead of
gold.