Adjust the vault muffindrake_ghost_three_little_ghosts
Rename this vault to muffindrake_ghost_three_little_ghosts to be in line
with other vault name formatting. Don't use diggable rock in the walls
in the vault to prevent monsters digging out/in. Instead use
stone/crystal/iron progression to make the theme consistent, and since
crystal has a nice interaction with the room 2 monsters. Remove outer
glass, since vault already reveals each rooms contents as the player
progresses.

Use a shaman band in the first room instead of just a single
non-threatening deep troll. Add randomization to the types and numbers
of monsters in the remaining two rooms. Increase loot to be more in line
with what we place for other Depths ghost vaults.