Adjust the vault muffindrake_ghost_dead_mans_drink
Rename this vault to muffindrake_ghost_dead_mans_drink to keep vault
naming consistent. Add "portholes" (transparent stone) to some sides of
the vault so that the ghost can be seen from outside even if it wanders
into the back of the vault.

Add randomization to the kinds of derived undead used
in this vault and randomize the placement and numbers of monsters,
keeping the same average. Increase the loot a little to keep numbers
more in line with other Depths ghost vaults.