A demon polymorph themed Depths ghost vault from Biasface
In this vault, the unfortunate player ghost hovers over a
(pre-identified) wand of polymorph, their final act having been an
attempt to polymorph a demon into something less threatening. Was it a
hellion? A tormentor?  Perhaps the player polymorphed a mere neqoxec to
avoid getting a **game-ending mutation**? Whatever their choice, the
untimely result was a tier 1 demon and, ultimately, a player ghost. The
vault also contains a few tier 2 demons (although no
tormentors/hellions) and a group of other common demons (tiers 3 and 4).

Changes from the original vault include a little vault surgery to make
the lava pool larger, more randomization to the composition of demons,
and randomization of the numbers and placement of demons. Loot is
adjusted to be roughly similar to what other ghost vaults give in
Depths. Also make the statue tiles a random assortment of demon-themed
ones.