Updates to ebering_funeral_grove
Some significant modifications to the monster set for this this vault.
Spriggan druids don't work as well in Dungeon and Lair, since Awaken
Forest requires it to be hard to get away from trees (i.e. be in Swamp)
and Druid's Call needs most level monsters to be beasts (not the case in
Dungeon). Rework the monster list, retaining the spriggans that always
have polearms, but adding in some wolves/bears and eventually a boggart
to replace the spiritual role of the druid. Additionally always give one
spriggan an artefact spear or trident both as part of the loot and to
highlight the troubles of an unfortunate player stuck in tree form
versus a monster with a nasty reach attack. Initially use only ordinary
spriggans but later introduce spriggan riders. In Swamp we can use the
spriggan druid again and use 098 Swamp monsters instead of wolves/bears.
The final average amounts of loots and monsters are similar to the
numbers before these changes.

Also add better windows into this vault so that any wandering ghost can
always be seen, and generally simplify glyph usage and SUBST/NSUBST.
Move the decoration statements towards the end of the vault definition
where we customarily put these.