Bug fixes and code generalizations for gammafunk_ghost_hdyra_chop
Generalize the random item definition code so that it can be used for a
forthcoming vault. Move the general parts to a new `random_item_def()`
function that allows specifying the item/ego weights as tables, with the
ego optional, as well as the separator to use in the definition. The
latter will allow is to use this in items in MONS statements. Simplify
the weighting used in this function when both items and egos are given
to simply multiply the weights together. A scaling factor is not
necessary if we compute the weights this way.

In the vault itself, fix the erroneous specification for long swords and
rename the function to use 'setup', since it's not initializing
anything.