An orcish armoury ghost vault for Dungeon and Orc (Biasface)
In biasface_ghost_orc_armoury, a player tried to raid an orcish armoury
and was overwhelmed by the defenses. Hopefully you can succeed where
they failed...just keep in mind that the player's ghost has joined
forces with the orcs in guarding the base.

The vault places melee orcs in substantial numbers, introducing a
smattering of orc priests after D:3. The ordinary orcs are reinforced
with warriors, knights, and a possible warlord in the latest depths. For
the top tier orc class used, one such monster will always have a fancy
weapon and armour (each either good_item or randart), and the additional
loot items are aux armour, general armour, weapon, and jewellery that
get progressively introduced and upgraded with good_item.

Biasface made the original vault submission. I made some slight layout
changes and added many of the item and monster progression details as
well as code support for setting equipment on the higher-level orcs.