More ghost vaults
More ghost vaults, from gammafunk and ebering, consisting of 10 commits:

 2af43bd Gozag themed ghost vault

No items, only gold and a Gozag altar. Gold loot is scattered in piles
and scales as follows: From 3-9 piles at D:3 to 12-18 at D:15 and below.
Further scaling comes from the natural scaling of gold pile size. The
monster set has two 998 monster placements and adds some 0's deeper.
There is a chance for the central statue to be an oklob in lair or an
obsidian statue elsewhere.

Committer's Note: Squashed two other subsequent commits by ebering and
updated the final commit message based on his notes in the squashed
commits. This vault's placement depth, loot, and monster composition
will be modified in a subsequent commit.

 bb431e2 Disaster themed lone ghost vault

 0289154 Davey Jones' / Kraken themed ghost vault for Shoals

Guarantees a randart ring, chance of an octopus king ring

 a6a7277 Port several ghost vaults to the Vaults branch

Committer's Note: Squashed commit that ports the lone ghost vaults
ebering_ghost_inner_flame, ebering_ghost_reflecting_pool, and
ebering_ghost_disaster_area to vault rooms for Vaults.

 a172949 Updates to ebering_ghost_gozag

Scale the threats more carefully so that it doesn't feature so many 9
and 8 monsters in earliest depths and so that difficulty increases more
smoothly with depth. Always place 2 normal random spawns on average,
slowly increasing the monster count by adding 9 and 8 monsters. Tweak
the placement of the obsidian statue to increase in chance gradually,
making it 100% by Lair branches. To supplement the loot as depth
increases, add a mix of either a good loot potion or an antique shop,
starting with a 50% chance of one of these on D:9, making up to two of
them by the later depths.

Simplify the logic of the gold pile function to mostly use absdepth,
increasing the value used a bit for Orc. Also add a variant of this
vault as a ghost room for Vaults.

 5148d60 Updates to ebering_ghost_davey_jones

Allow this vault to place on Shoals:$, since it will work fine on this
level. Instead of either a kraken or a kraken simulacrum, place either a
single kraken or a mix of two kraken simulacrum or spectres, as either
of the latter have vastly less HP than a kraken. Add a bit more deep
water to give the pair more room to move about.

Refactor glyph usage to not use the % glyph for corpses and hence fix a
loot item.

 afc7989 Some ghost vault DES and code cleanups and refactoring

Use the general loot functions for lone ghost vaults in the Vaults ghost
rooms, adding it to ghost.lua so it's available to those vaults. Trim
extra outer spacing or wall in some of these same ghost rooms and remove
any special floor tile settings. These tiles look out of place in Vaults
rather than conveying a ruined look. Don't try to use sandstone wall
tiles in the Gozag ghost vaults since it conveys more of a weathered
natural look than it does "gold". Likewise remove the yellow stone wall
coloring for console.

Move several ghost vaults that have thematic monsters into the "Themed"
section of ghost.des, add a missing float ORIENT, and fix a comment.

 b62a27e Port some ghost vaults to the Vaults branch

Adapt ebering_ghost_xom, biasface_ghost_orc_armoury, and
ploomutoo_gammafunk_potion_laboratory as ghost rooms for Vaults.
Ebering's vault is copied basically unchanged, just with some extra
space to not make the room too tiny. For Biasface's vault, I've removed
the wargs, orc priests, and some of the plain orcs, replacing them with
orc high priests, some more orc knights, stone giants, and orc warlords.
The armour/weapon/jewellery loot has one piece of good potion or scroll,
and the additional warlords that sometimes place have a chance to get
fancy gear. Ploomutoo's vault is ported mostly unchanged, using the same
monster and loot set that the original vault uses in Depths, but with
slightly fewer top-tier monsters and one less loot piece in each
chamber.

 65f81ef Updates to gammafunk_ghost_icy

Tweak the numbers of monsters and loot items to have a more consistent
progression and fix floor tiles surrounding the vault.

 46e8f3c A stat drain themed ghost vault for Dungeon and Orc (mikee)

Players who don't watch out for those pesky sources of stat drain might
end up haunting the transporter vault gammafunk_ghost_stat_death.
Dex-draining Quasits are the only threat accompanying the ghost early
on, with the quasits being supplemented by int-draining neqoxecs by D:9.
There's the usual pile of progressive loot, and the ghost generates over
a ring of dexterity or intelligence that has progressively higher
chances for being good_item or randart.

Based on an initial suggestion from mikee for a ghost vault with a bunch
of quasits and a ring of DEX.