And icy themed ghost vault for Dungeon, Lair, and Depths (Biasface)
An icy cave filled with monsters with cold melee and ranged attacks. The
ghost spawns over a pre-type-identified ring of fire; sadly the fallen
player did not remember to remove their source of rF- in time. The ring
of fire eventually has a chance to be an artifact, with that chance
reaching 100% in the latest depths. The other loot is typical for other
ghost vaults, a progressive mix of superb_item and star_item. The
composition and number of icy monsters increases in difficulty with
depth.

Based on a vault idea submitted by Biasface.