Move a function to the ghost vault lua file and refactor
The random_item_def() function will likely be of use to vaults e.g. in
Vaults, so move this to ghost.lua and add documentation. Generalize the
"artefact" argument to just be an argument string, allowing callers to
set whatever additional item properties they want instead of just
good_item or randart.