Add good consumables or equipment to the loot in many ghost vaults
Currently, ghost vaults that place monsters in addition to the ghost
also place items that upgrade in quality, with the count increasing from
3 to as high as 7 as depth increases. These items are of higher quality
than usual floor items, with each item eventually reaching 50% each of
superb_item and star_item, but those item classes are highly variable.
In later Dungeon levels and beyond, it's common to see no item that's
reasonably worth the risk of fighting a ghost along with additional
monsters. To provide additional incentive for players to attempt these
vaults, this commit adds a `good_ghost_loot()` function that can replace
one or two pieces of the "superb_item / star_item" loot with items that
have a narrower range of useful possibilities.

The first replacement consists of scrolls and potions with weights only
for those that are broadly useful, along with aux armour, jewellery, and
a small chance for a manual. The aux armour and jewellery have lower
weight relative to the scrolls and potions, but upgrade to have
good_item and later randart modifiers as depth increases. The first item
replacement only has a chance to happen starting from D:9-11, with the
chance increasing to 100% by Lair. The second item is a potion or scroll
using the same weights as the first item. It has a chance to appear
starting in Lair, becoming 100% by the Lair branches.

Most ghost vaults that place additional monsters have been updated to
call this function and have any necessary change to their loot placement
statements. The function defines the items on the 'd' and 'e' glyphs and
leaves the calling vault to perform any necessary substitutions if they
wish to override this definition or handle multiple instances of those
glyphs. A couple vaults that already place specialized loot remain
unchanged.