A labyrinth themed ghost vault for Dungeon, Lair, and Orc from Maelrawn
In maelrawn_ghost_labyrinth_escape, a player tried to escape from a lab
with loot in tow. Sadly, the minotaur was right behind them, and then
they ran right back into all the monsters they ditched to reach the lab
entrance.

This vault places on any level that could normally place a labyrinth
(D:10-, Lair, Orc:2), but of course only places through the usual ghost
vault chance and additionally has half the usual vault weight. It should
be a pretty nasty one, since the minotaur is quite the threat on its
own. The vault supplements the minotaur with a few branch monsters that
eventually include out-of-depth 9 monsters at later placement depths. It
doesn't recreate the more generous loot of an actual Labyrinth portal,
but it does place up to 5 items and supplements this with the good loot
replacement used by other ghost vaults.