Revise the mace of Variability
This is a squashed version of 6 commits from ebering and gammafunk which
rework the mace of Variability to cast chain of chaos that doesn't hurt
the wielder.

 8c3e449 Revise the mace of Variability

Having a weapon that was occasionally useless didn't make for a good
item. In the tradition of other unrands that are "this brand but more"
(Arc Blade, Firestarter, Frostbite, Leech), this changes the mace to
have a chance of casting chain of chaos on hit and have a fixed plus. As
flavourful as the plus swing is the player is encouraged to wait for the
random walk to go up before fighting.

Committer's Note: The chain of chaos damage and set of effects as well
as the enchantment for this weapon are modified in a subsequent commit.

 9148847 Add the mace of Variability to ebering_ghost_xom

The item is terrifying now that it has chain of chaos, so it can't
appear early. But also thematic and good temptation.

 0c8d57c Adjust effect damage and enchantment of the mace of variability

The mace of variability casts chain of chaos with spell-power ranging
from 75 to 149 with an average of 112. Ignoring any damage from the
chaos beam effect, the current arc damage formula of 3d(5 + pow / 2)
gives the mace's spell damage die of 3d61 per arc on average. This is
very high, especially since the monster gets zapped 4 times on average
(albeit with spell-power that decreases by 14 on average with each arc).

To balance the damage from the mace, reduce the damage formula of chain
of chaos to 3d(5 + pow / 6). Now an arc from the spell will do 3d23
damage on average, which is still significant damage even when cast only
1 in 5 melee hits. And of course, in addition to damage, each arc
applies random effect from the chaos beam.

This commit also lowers the enchantment of the mace to +7 from a
previous commit's value of +10. A +7 great mace will be quite strong
enough with 1 in 5 attacks casting chain of chaos in addition to the
normal chaos ego. Also 7 is a lucky number, which Xom appreciates, and a
value lower than the typical +9 max is a **secret reference** to the
original unrand behaviour.

 f487629 Adjust the effects of chaos beams

The chaos beam type is only used for the chain of chaos spell cast by
the mace of variability and for the impact (non-cloud) effect of
apocalypse crab breath. Currently chaos beams select from a set of beams
analogous to effects used by the chaos brand, but using mostly flat
selection weights. The flat weights effectively give chaos beams much
higher weighting relative to the chaos brand for strong effects like
banishment, polymorph, hexing or buffing the monster, or miscast
effects. This difference isn't very apparent for chaos breath, since the
beam is only applied once for impact damage, but for the mace of
variability, a chaos arc can hit a monster multiple times in a turn.

This commit re-weights the chaos beam effect table to be closer in
behaviour to the chaos brand. Beams directly analogous to other brands
like flaming, freezing, electrocution, etc now have comparatively higher
weight, and rarer effects that only happen when the chaos brand
reselects itself have lower weight.

Note that this commit has no effect on the behaviour of the chaos brand,
monster af_chaos, or chaos clouds; it's only relevant for the mace of
variability and impact (non-cloud) effects from chaos breath. It would
be ideal to somehow merge the selection of these beam effects with the
effect selection used for SPWPN_CHAOS, especially since the latter is
already used for all chaos melee and launcher egos as well as chaos
clouds.

 e3ae32f Remove unused monster casting for chain of chaos

This spell used to be castable on a demonspawn monster, but that monster
no longer exists. Chain of chaos isn't very good as a monster spell
since it applies multiple chaos effects to the player per turn and at
range. Chaos breath is ranged but only applies on effect per turn due to
its use of clouds, so this is at least a better style of spell or breath
to give monsters.

This commit removes code used to let monsters cast chain of chaos
directly. Monsters can still use the chain of chaos effect from wielding
the mace of variability, since that triggers from a melee effect and
doesn't need the usual monster spell setup functions.

 7d32d7c Update the description of the mace of variability

Don't refer to the chain of chaos spell, which isn't directly castable.
Instead refer to the effects right in the description. Also change the
flavor text to be more inline with the current weapon behavior.

Closes #802.