Remove self-harm from chain of chaos (gammafunk)
The effects (even with a special case preventing banishment and
polymorph) didn't make for a good monster spell because of self-harm. This
adds a special case to prevent chain of chaos from harming the caster,
replacing the beam's flavour with BEAM_VISUAL to keep the arcs going and
waste an arc so that the spell has about the same offensive power as
before. Banishment and polymorph are now added to the effect list.