Updates to gammafunk_ghost_split
For later depths, update the monsters this vault places to use a bit
less 0 monsters and a little more of 9 and 8 out-of-depth ones. Early on
it's ok if this vault feels less like a 098 vault and more like a vault
with in-depth monsters only, but by Lair branches it should really have
a significant number of more threatening monsters in addition to the
ghost. Also update the wall glyph usage so that the outer or inner walls
would use wall types that would seem out of place in certain branches.