More ghost vaults
 b275bf3 Some ghost vault fixes and refactoring.:

Make the good loot potion/scroll item lists available as variables for
other vaults to use when they're handling their own loot placement. Fix
a potion specification error ('cancellation', not 'cancelation') and
clean up DES spacing.

 ff9aefc A potion ghost vault for Dungeon, Elf, and Depths (Ploomutoo):

In ploomutoo_gammafunk_potion_laboratory, an alchemist with a collection
of nice potions conducted a fatal experiment resulting in an explosion
of mutagenic horrors. Now their ghost haunts a two-chambered lab,
guarding loot that contains a number of potions of mutation, high
quality random potions, and a piece of aux armour or a robe. Monsters in
the first chamber are melee-oriented: ugly things, very ugly things,
large abominations, and tentacled monstrosities are all used depending
on the depth and branch, supplemented with the occasional glowing orange
brain. In the second chamber is the ghost along with shape-shifters and
malmutators (neqoxecs and later cacodemons).

This vault was originally submitted by ploomutoo with the main layout
and a basic set of monsters and loot for dungeon. I modified the monster
sets to make the themes of the chambers distinct and to allow placement
in other branches. I also scaled the number and composition of loot and
monsters as depth increases. Finally, the layout was updated to better
allow placement of mutagenic fog machines.

 270d1ae A happy afterlife ghost vault for Dungeon and Lair from hellmonk:

In hellmonk_ghost_happy_afterlife, a ghost decides to not dwell on the
failures they made in life, preferring to just relax next to a nice tree
with a nice meal (a ration) and some nice drinks (1-2 potions). Of
course, if you disturb its meal it will not hesitate to kill you, since
ghosts like killing players more than they like meals they can no longer
eat. The potions upgrade in quality and number as depth increases.