A potion ghost vault for Dungeon, Elf, and Depths (Ploomutoo)
In ploomutoo_gammafunk_potion_laboratory, an alchemist with a collection
of nice potions conducted a fatal experiment resulting in an explosion
of mutagenic horrors. Now their ghost haunts a two-chambered lab,
guarding loot that contains a number of potions of mutation, high
quality random potions, and a piece of aux armour or a robe. Monsters in
the first chamber are melee-oriented: ugly things, very ugly things,
large abominations, and tentacled monstrosities are all used depending
on the depth and branch, supplemented with the occasional glowing orange
brain. In the second chamber is the ghost along with shape-shifters and
malmutators (neqoxecs and later cacodemons).

This vault was originally submitted by ploomutoo with the main layout
and a basic set of monsters and loot for dungeon. I modified the monster
sets to make the themes of the chambers distinct and to allow placement
in other branches. I also scaled the number and composition of loot and
monsters as depth increases. Finally, the layout was updated to better
allow placement of mutagenic fog machines.