Merge deep elf and tengu into cloud elf
This is the initial commit for merging DE and Te into Cl (we require the
unusual abbreviation Cl to avoid confusion with Ce, centaur).

Currently they have aptitudes which are half way between those of DE and
Te.

There will be more commits following, for adjustments and other code
improvement reasons.