Exploration based trap effects.
Each tile revealed has a small chance (the chance of generating a shaft
trap on that tile for that particular level) of shafting or alarm
trapping (thanks gammafunk) the player.

With all traps known the 'interesting situation' aspect of shaft (and
other) traps was lost. This random explore effect recovers that aspect.

It does mean that a player wishing to avoid traps should avoid revealing
tiles. This is a better trade-off than the previous behavior.  While the
'optimal' play is to not reveal more tiles than necessary, the act of
revealing (instead of stepping on) has the potential benefit of giving
an item or XP from an easy monster. Further, unlike tracking stepped on
tiles, the UI makes it possible to reason about what moves will reveal
tiles; and waypoints provide a method of charting particular routes
through partially explored areas.

However, as minmay has pointed out, Gnolls, Xom, and Ash all give item
detection which allow a player to avoid revealing tiles that they know
will give no benefit. While minmay presents this as a problem, I see it
as a benefit of item detection.