Make items for the cloak slot be equally common as other aux slots again
In commit 651d724, scarves were increased from being 1/6 to 1/4 of items
generated for the cloak slot, however the actual increase in spawn rate
for scarves was minimal, from 1/30 to 1/28 of all auxiliary slot armours
generated, because what this commit actually did was reduce the spawn
rate of cloaks from 5/30 to 3/28 by reducing the weight of cloaks.

It seems like an accident that this meant the total weight for this
random_choose_weighted was not the 30 that a comment implied, and that
the cloak slot items (cloaks and scarves combined) were made 2/3 as
common as other auxiliary slot armours.

As a result, this commit buffs the spawn rate of cloaks and scarves to
bring the overall spawn rate in line with that of other auxiliary
armours, whilst maintaining the same 3:1 ratio of cloaks:scarves.