Updates to ebering_ghost_gozag
Scale the threats more carefully so that it doesn't feature so many 9
and 8 monsters in earliest depths and so that difficulty increases more
smoothly with depth. Always place 2 normal random spawns on average,
slowly increasing the monster count by adding 9 and 8 monsters. Tweak
the placement of the obsidian statue to increase in chance gradually,
making it 100% by Lair branches. To supplement the loot as depth
increases, add a mix of either a good loot potion or an antique shop,
starting with a 50% chance of one of these on D:9, making up to two of
them by the later depths.

Simplify the logic of the gold pile function to mostly use absdepth,
increasing the value used a bit for Orc. Also add a variant of this
vault as a ghost room for Vaults.