Updates to ebering_ghost_davey_jones
Allow this vault to place on Shoals:$, since it will work fine on this
level. Instead of either a kraken or a kraken simulacrum, place either a
single kraken or a mix of two kraken simulacrum or spectres, as either
of the latter have vastly less HP than a kraken. Add a bit more deep
water to give the pair more room to move about.

Refactor glyph usage to not use the % glyph for corpses and hence fix a
loot item.