Rework Dithmenos' Shadow Step restrictions and costs (mikee)
Shadow Step currently requires that any target monster be "still" in
order for it to be a valid target, which means the monster has to be
asleep, paralyzed, or otherwise immobile. This generally requires using
the ability near unexplored areas, which is risky even if the player can
one-shot the target monster with a stab. Players without good stabbing
ability found little use for Shadow Step despite the ability being the
only relatively low cost invocable offered.

This commit removes the stillness restriction for Shadow Step, instead
giving the ability an HP cost of about 8% of max HP. This will give
incentive for non-stabbers to use the ability while still being useful
for setting up stabs for high-priority targets. Shadow Step still
requires the target to be within the Umbra, but no longer requires any
notion of monster stillness.

Balance-wise, allowing the player to make long chains of shadow step is
a concern, but a significant HP cost paired with the current fairly high
piety cost prevent this from getting out of hand. As the player loses
piety, their access to Shadow Form can be lost and the size of the Umbra
decreases, reducing the range of the next shadow step. We can adjust the
HP and piety costs accordingly if the current costs aren't sufficient.

The ability description has been updated to remove references to
stillness and to clarify that the adjacent landing site is randomly
chosen.