Make Dithmenos an evil god (mikee)
Now that Shadow Step has an HP cost, this reinforces the notion that
Dithmenos is an evil god. It wants the entire world to be shrouded in
eternal darkness; doesn't that sound just mean and awful? The gameplay
consequences of this are few, mostly related to conversion wrath from
the good gods, the god-themed artefacts not getting the holy wrath
brand, and the inability to summon or phantom mirror holy monsters. It
will give Dith a bit more flavor to replace that which was lost when the
fire conduct was removed.

The new god description text, which is surprisingly non-transifex-like,
comes courtesy of mikee.